Reduce the amount of vector load instructions (from "12 vbroadcastsd + 2 vmovupd" to "6 vbroadcastf32x4 + 4 vmovddup" per iteration) to minimize power consumption and thermal throttling.

Performance of DGEMM with 26 threads on Intel Xeon Platinum 8269CY (transa = transb = N):

Dimension | GFLOPS-OpenBLAS-old | GFLOPS-OpenBLAS-new | GFLOPS-MKL2019
-- | -- | -- | --
10000 | 1348 | 1396 | 1527
20000 | 1401 | 1443 | 1551
30000 | 1400 | 1449 | 1555