@jeffhammond 2020-09-03
Have you measured real-time frequency data to prove that this change has any impact on power or thermals?

I don't think there is any basis to the suggest that.  Rather, you have optimized data movement and that improves performance.  This also has the effect of increasing the frequency of AVX512 FMA instructions, which - if anything - will increase power consumption, since FMAs are AVX512-heavy while load-store are AVX512-light.  However, given the overall intensity of FMAs in DGEMM, I would expect this change has no effect on frequency, because the timescale for frequency changes is much greater than that of load-store in the inner loop of DGEMM.

I work for Intel and consulted with our AVX512 DGEMM experts before posting this.  I would like this issue to be accurate because it has already convinced someone else that throttling, not reducing loads and stores, is the cause for the performance change (https://github.com/flame/blis/issues/441).